requirements.txt
flask==2.0.1
pyyaml==6.0
requests==2.28.1
prometheus_client==0.17.0
openai-whisper==20231117
werkzeug==2.0.3
gunicorn==20.1.0